Created separate files with specs for each object.
