--- spec/tic-tac-toe.spec.js
@@ -1,3 +1,0 @@
-import Game from 'tic-tac-toe';
-import Player from 'tic-tac-toe';
-import Board from 'tic-tac-toe';
