+++ src/board.js
@@ -1,0 +1,4 @@
+var Board = function() {};
+
+
+export default Board;
