+++ src/game.js
@@ -1,0 +1,6 @@
+var Game = function() {};
+
+
+
+
+export default Game;
