+++ src/player.js
@@ -1,0 +1,8 @@
+var Player = function() {};
+
+
+
+
+
+
+export default Player;
