--- src/tic-tac-toe.js
@@ -1,11 +1,0 @@
-var Game = function() {};
-var Player = function() {};
-var Board = function() {};
-
-
-
-
-
-export default Game;
-export default Player;
-export default Board;
